Response from API service, adapter created
