+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="movies.flag.pt.moviesapp">
+          package="topRatedMovies.flag.pt.moviesapp">
 
     <uses-permission android:name="android.permission.INTERNET"/>
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
@@ -13,19 +13,19 @@
 
         <!-- Activities Declaration -->
 
-        <activity android:name=".screens.StartScreen">
+        <activity android:name="topRatedMovies.flag.pt.moviesapp.screens.StartScreen">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
 
-        <activity android:name=".screens.LatestMovieScreen"/>
-        <activity android:name=".screens.PopularTvShowScreen"/>
-        <activity android:name=".screens.DetailMovieScreen"/>
-        <activity android:name=".screens.SearchMovieScreen"/>
-        <activity android:name=".screens.DetailTvShowScreen"/>
-        <activity android:name=".screens.RandomMovieScreen"/>
+        <activity android:name="topRatedMovies.flag.pt.moviesapp.screens.LatestMovieScreen"/>
+        <activity android:name="topRatedMovies.flag.pt.moviesapp.screens.PopularTvShowScreen"/>
+        <activity android:name="topRatedMovies.flag.pt.moviesapp.screens.DetailMovieScreen"/>
+        <activity android:name="topRatedMovies.flag.pt.moviesapp.screens.SearchMovieScreen"/>
+        <activity android:name="topRatedMovies.flag.pt.moviesapp.screens.DetailTvShowScreen"/>
+        <activity android:name="topRatedMovies.flag.pt.moviesapp.screens.RandomMovieScreen"/>
 
 
         <!-- Database configurations -->
