--- app/src/main/java/movies/flag/pt/moviesapp/http/entities/Dates.java
@@ -1,53 +1,0 @@
-
-package movies.flag.pt.moviesapp.http.entities;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-
-public class Dates {
-
-    @SerializedName("maximum")
-    @Expose
-    private String maximum;
-    @SerializedName("minimum")
-    @Expose
-    private String minimum;
-
-    /**
-     * 
-     * @return
-     *     The maximum
-     */
-    public String getMaximum() {
-        return maximum;
-    }
-
-    /**
-     * 
-     * @param maximum
-     *     The maximum
-     */
-    public void setMaximum(String maximum) {
-        this.maximum = maximum;
-    }
-
-    /**
-     * 
-     * @return
-     *     The minimum
-     */
-    public String getMinimum() {
-        return minimum;
-    }
-
-    /**
-     * 
-     * @param minimum
-     *     The minimum
-     */
-    public void setMinimum(String minimum) {
-        this.minimum = minimum;
-    }
-
-}
