--- app/src/main/java/movies/flag/pt/moviesapp/http/entities/PopularTvShowResponse.java
@@ -1,58 +1,0 @@
-package movies.flag.pt.moviesapp.http.entities;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-/**
- * Created by Marina on 25/01/2017.
- */
-
-public class PopularTvShowResponse {
-
-    @SerializedName("page")
-    @Expose
-    private Integer page;
-    @SerializedName("results")
-    @Expose
-    private List<ResultPopularTvShow> results = null;
-    @SerializedName("total_results")
-    @Expose
-    private Integer totalResults;
-    @SerializedName("total_pages")
-    @Expose
-    private Integer totalPages;
-
-    public Integer getPage() {
-        return page;
-    }
-
-    public void setPage(Integer page) {
-        this.page = page;
-    }
-
-    public List<ResultPopularTvShow> getResults() {
-        return results;
-    }
-
-    public void setResults(List<ResultPopularTvShow> results) {
-        this.results = results;
-    }
-
-    public Integer getTotalResults() {
-        return totalResults;
-    }
-
-    public void setTotalResults(Integer totalResults) {
-        this.totalResults = totalResults;
-    }
-
-    public Integer getTotalPages() {
-        return totalPages;
-    }
-
-    public void setTotalPages(Integer totalPages) {
-        this.totalPages = totalPages;
-    }
-}
