--- app/src/main/java/movies/flag/pt/moviesapp/http/requests/GetNowPopularTvShowAsyncTask.java
@@ -1,34 +1,0 @@
-package movies.flag.pt.moviesapp.http.requests;
-
-import android.content.Context;
-
-import movies.flag.pt.moviesapp.http.entities.PopularTvShowResponse;
-
-/**
- * Created by Marina on 25/01/2017.
- */
-
-public abstract class GetNowPopularTvShowAsyncTask extends ExecuteRequestAsyncTask<PopularTvShowResponse> {
-
-    private static final String PATH = "/tv/popular";
-
-
-    public GetNowPopularTvShowAsyncTask(Context context) {
-        super(context);
-    }
-
-    @Override
-    protected String getPath() {
-        return PATH;
-    }
-
-    @Override
-    protected void addQueryParams(StringBuilder sb) {
-
-    }
-
-    @Override
-    protected Class<PopularTvShowResponse> getResponseEntityClass() {
-        return PopularTvShowResponse.class;
-    }
-}
