+++ app/src/main/java/topRatedMovies/flag/pt/moviesapp/http/entities/TopRatedMovieResponse.java
@@ -1,0 +1,56 @@
+
+package topRatedMovies.flag.pt.moviesapp.http.entities;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+import java.util.List;
+
+public class TopRatedMovieResponse {
+
+    @SerializedName("page")
+    @Expose
+    private Integer page;
+    @SerializedName("resultTopRatedMovies")
+    @Expose
+    private List<ResultTopRatedMovie> resultTopRatedMovies = null;
+    @SerializedName("total_results")
+    @Expose
+    private Integer totalResults;
+    @SerializedName("total_pages")
+    @Expose
+    private Integer totalPages;
+
+    public Integer getPage() {
+        return page;
+    }
+
+    public void setPage(Integer page) {
+        this.page = page;
+    }
+
+    public List<ResultTopRatedMovie> getResultTopRatedMovies() {
+        return resultTopRatedMovies;
+    }
+
+    public void setResultTopRatedMovies(List<ResultTopRatedMovie> resultTopRatedMovies) {
+        this.resultTopRatedMovies = resultTopRatedMovies;
+    }
+
+    public Integer getTotalResults() {
+        return totalResults;
+    }
+
+    public void setTotalResults(Integer totalResults) {
+        this.totalResults = totalResults;
+    }
+
+    public Integer getTotalPages() {
+        return totalPages;
+    }
+
+    public void setTotalPages(Integer totalPages) {
+        this.totalPages = totalPages;
+    }
+
+}
