+++ app/src/main/java/topRatedMovies/flag/pt/moviesapp/utils/ResponseErrors.java
@@ -1,0 +1,14 @@
+package topRatedMovies.flag.pt.moviesapp.utils;
+
+/**
+ * Created by Ricardo Neves on 20/01/2017.
+ */
+
+public class ResponseErrors {
+
+    public static final int NO_ERROR = 0;
+    public static final int NETWORK_ERROR = 1;
+    public static final int PARSE_ERROR = 2;
+    public static final int URL_MALFORMED = 3;
+
+}
